refactor: move Pragma to ./state and delete ./adapter
