+++ packages/react/tsup.config.ts
@@ -6,10 +6,10 @@ export default defineConfig([
     sourcemap: true,
     clean: true,
     outDir: 'dist',
-    external: ['./adapter', './jsx-runtime'],
+    external: ['./state', './jsx-runtime'],
     entry: {
       index: 'src/index.ts',
-      adapter: 'src/adapter.ts',
+      state: 'src/state.ts',
       'jsx-runtime': 'src/jsx-runtime.ts',
       'jsx-dev-runtime': 'src/jsx-dev-runtime.ts'
     },
@@ -18,11 +18,11 @@ export default defineConfig([
   {
     sourcemap: true,
     outDir: 'dist/esm',
-    external: ['./adapter', './jsx-runtime'],
+    external: ['./state', './jsx-runtime'],
     outExtension: () => ({ js: '.js' }),
     entry: {
       index: 'src/index.ts',
-      adapter: 'src/adapter.ts',
+      state: 'src/state.ts',
       'jsx-runtime': 'src/jsx-runtime.ts',
       'jsx-dev-runtime': 'src/jsx-dev-runtime.ts'
     },
